Adding a writer for Markua 0.10
This change will add a simple writer to support outputs in Markua 0.10. The
writer setup is similarly to the Markdown writer to conform to the pandoc
structure.

Integrate markua writer to markdown writer

According to the feedback on our initial [pull request](https://github.com/jgm/pandoc/pull/7729) we integrated the writer with the markdown writer.

correct and improve Markua parts according to feedback